# Use single allocation for indirect values in array_multisort